@@ -35,7 +35,7 @@ dependencies:
   turf: ^0.0.10
   logger: ^2.4.0
   postgrest: ^2.3.0
-  mapbox_maps_flutter: ^2.3.0
+  mapbox_maps_flutter: ^2.5.0
   postgres: ^3.4.0
   provider: ^6.1.2
 
@@ -43,6 +43,9 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.8
+  overlay_support: ^2.1.0
+  intl: ^0.20.1
+  another_flushbar: ^1.12.30
 
 dev_dependencies:
   flutter_test:
